+++ database/migrations/2026_02_03_203401_create_producto_proforma_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('producto_id')->constrained('productos')->onDelete('cascade');
             $table->foreignId('proforma_id')->constrained('proformas')->onDelete('cascade');
             //descuento
-            $table->float('descuento_cliente', 5, 2)->default(0);
+            $table->decimal('descuento_cliente', 5, 2)->default(0);
             $table->integer('cantidad')->default(1);
             $table->decimal('precio_unitario', 10, 2);
             $table->timestamps();
